DenseNet/latest/requirements.txt
numpy==1.21.6
Pillow==9.5.0
Pillow==10.0.0
torch==1.9.0+cu111
torchvision==0.10.0+cu111
tqdm==4.66.1
